+++ migrations/mysql/2026-07-26-000004_add_users_organizations_paging_index/down.sql
@@ -1,1 +1,0 @@
-DROP INDEX idx_users_organizations_org_uuid_paged ON users_organizations;