+++ migrations/postgresql/2026-07-26-000005_add_groups_paging_index/down.sql
@@ -1,1 +1,0 @@
-DROP INDEX IF EXISTS idx_groups_org_uuid_paged;