+++ migrations/sqlite/2026-07-26-000003_add_scim_api_key_last_used/down.sql
@@ -1,1 +1,0 @@
-ALTER TABLE scim_api_key DROP COLUMN last_used_at;